refactor: 컴포넌트 오타 변경
